IHCIC-76 Adding the connect method. Users of RestConnection will have to call connect now.
